# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Landa Travels conversational planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assistant/handover/status:
    post:
      operationId: getGuestHandoverStatus
      tags: [assistant]
      summary: Read guest request status using its private conversation capability
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GuestHandoverStatusInput" }
      responses:
        "200":
          description: Request status and any advisor reply; no contact email or transcript
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "404": { description: Request unavailable or conversation expired }
...
  /advisor/handovers:
    get:
      operationId: listAdvisorHandovers
      tags: [advisor]
      summary: List advisor handovers for authorized advisors
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Advisor inbox
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdvisorHandover" }
        "401": { description: Authentication required }
        "403": { description: Advisor access required }
  /advisor/handovers/{id}:
    get:
      operationId: getAdvisorHandover
      tags: [advisor]
      summary: Get the transcript and privacy-minimised trip context for the assigned advisor
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assigned advisor handover detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandoverDetail" }
        "401": { description: Authentication required }
        "403": { description: Advisor access required }
        "404": { description: Handover not found or not assigned to this advisor }
  /advisor/handovers/{id}/assign:
    post:
      operationId: assignAdvisorHandover
      tags: [advisor]
      summary: Assign a handover to the current advisor
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdvisorAssignmentInput" }
      responses:
        "200":
          description: Assigned handover
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "401": { description: Authentication required }
        "403": { description: Advisor access required }
        "404": { description: Handover not found }
        "409": { description: Invalid lifecycle transition or assignee conflict }
  /advisor/handovers/{id}/reply:
    post:
      operationId: replyToAdvisorHandover
      tags: [advisor]
      summary: Reply to a traveller handover
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdvisorReplyInput" }
      responses:
        "200":
          description: Replied handover
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "400": { description: Invalid reply }
        "401": { description: Authentication required }
        "403": { description: Advisor access required }
        "404": { description: Handover not found }
        "409": { description: Invalid lifecycle transition or assignee conflict }
  /advisor/handovers/{id}/reply-email:
    post:
      operationId: resendAdvisorReplyEmail
      tags: [advisor]
      security: [{ bearerAuth: [] }]
      summary: Retry emailing the assigned advisor's saved guest reply
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Saved reply with email-provider acceptance status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "404": { description: Guest reply not assigned to this advisor }
  /advisor/handovers/{id}/close:
    post:
      operationId: closeAdvisorHandover
      tags: [advisor]
      summary: Close a handover
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Closed handover
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdvisorHandover" }
        "401": { description: Authentication required }
        "403": { description: Advisor access required }
        "404": { description: Handover not found }
        "409": { description: Invalid lifecycle transition or assignee conflict }
...
components:
...
  schemas:
...
    CertificationEvidence:
      type: object
      additionalProperties: false
      required: [scheme, licenceNumber, validUntil, checkedAt, registryUrl, hotelUrl, address]
      properties:
        scheme: { type: string }
        licenceNumber: { type: string }
        validUntil: { type: string, format: date }
        checkedAt: { type: string, format: date-time }
        registryUrl: { type: string, format: uri }
        hotelUrl: { type: ["string", "null"], format: uri }
        address: { type: string }
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [stay, transport, experience, offset]
        name:
          type: string
        location:
          type: string
        description:
          type: string
        price:
          type: string
        carbonKg:
          type: ["number", "null"]
        carbonLabel:
          type: ["string", "null"]
          enum: [low, moderate, high]
        certificationEvidence:
          oneOf:
            - { type: "null" }
            - $ref: "#/components/schemas/CertificationEvidence"
        score:
          type: number
        durationMinutes:
          type: number
          minimum: 0
          description: Provider-supplied journey duration in minutes; absent when unknown. For return flight offers, sum of slice durations, not door-to-door time.
        connectionCount:
          type: integer
          minimum: 0
          description: Provider-supplied total connection count; absent when unknown.
        rankingExplanation:
          type: string
          description: Comparative weighting, known constraints and missing convenience evidence.
        certification:
          type: [string, "null"]
        source:
          type: string
        verifiedAt:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [id, type, name, location, description, price, carbonKg, carbonLabel, score, certification, source, verifiedAt, tags]
...
    GuestHandoverStatusInput:
      type: object
      required: [sessionId, handoverId]
      properties:
        sessionId: { type: string, minLength: 8, maxLength: 100 }
        handoverId: { type: string, minLength: 1, maxLength: 40 }
    HandoverTranscriptTurn:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 4000
    AdvisorHandoverContext:
      type: object
      required: [origin, destination, dateRange, travellerCount, budget, transportPreferences, accommodationNeeds, sustainabilityPriority, stopovers, accessibilityNeeds, locationConsentMode]
      properties:
        origin: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        dateRange: { type: ["string", "null"] }
        travellerCount: { type: ["integer", "null"] }
        budget: { type: ["string", "null"] }
        transportPreferences: { type: array, items: { type: string } }
        accommodationNeeds: { type: array, items: { type: string } }
        activityPreferences: { type: array, items: { type: string } }
        sustainabilityPriority: { type: ["string", "null"] }
        stopovers: { type: array, items: { type: string } }
        accessibilityNeeds: { type: array, items: { type: string } }
        locationConsentMode: { type: ["string", "null"] }
        recommendationOptions:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
        selectedRecommendationIds:
          type: array
          items: { type: string }
        recommendationSelectionStatus:
          type: string
          enum: [explicit_selection, no_explicit_selection]
        recommendationSource:
          type: ["string", "null"]
          enum: [demo, live, null]
        recommendationConfirmedAt:
          type: ["string", "null"]
          format: date-time
        recommendationUnknowns:
          type: array
          items: { type: string }
    AdvisorHandoverDetail:
      type: object
      required: [handover, context, transcript]
      properties:
        contactEmail: { type: ["string", "null"], description: Reply address, visible only to the assigned advisor. }
        handover: { $ref: "#/components/schemas/AdvisorHandover" }
        context: { $ref: "#/components/schemas/AdvisorHandoverContext" }
        transcript:
          type: array
          items: { $ref: "#/components/schemas/HandoverTranscriptTurn" }
...
    AdvisorHandover:
      type: object
      required: [handoverId, savedTripId, status, summary, travellerReply, assignedAdvisorId, assignedAdvisorName, notificationStatus, createdAt, updatedAt]
      properties:
        handoverId: { type: string }
        savedTripId: { type: ["integer", "null"] }
        replyNotificationStatus:
          type: string
          enum: [not_requested, pending, accepted, failed, not_configured]
        status:
          type: string
          enum: [requested, assigned, replied, closed]
        summary: { type: string }
        travellerReply: { type: ["string", "null"] }
        assignedAdvisorId: { type: ["string", "null"] }
        assignedAdvisorName: { type: ["string", "null"] }
        notificationStatus:
          type: string
          enum: [accepted, pending, failed, not_configured]
        createdAt: { type: string, format: date-time }
        assignedAt: { type: ["string", "null"], format: date-time }
        repliedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    AdvisorAssignmentInput:
      type: object
      properties:
        advisorName: { type: ["string", "null"], maxLength: 120 }
    AdvisorReplyInput:
      type: object
      required: [reply]
      properties:
        reply: { type: string, minLength: 1, maxLength: 4000 }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer